+++ R/show_sig_bootstrap.R
@@ -23,7 +23,10 @@
 #' @param signatures signatures to show.
 #' @param measure measure to estimate the exposure instability, can be one of 'MRSE', 'MAE' and 'AbsDiff'.
 #' @param dodge_width dodge width.
-#' @param ... other parameters passing to [ggpubr::ggboxplot].
+#' @param plot_fun set the plot function.
+#' @param highlight set the color for optimal solution. Default is "auto", which use the same color as
+#' bootstrap results, you can set it to color like "red", "gold", etc.
+#' @param ... other parameters passing to [ggpubr::ggboxplot] or [ggpubr::ggviolin].
 #'
 #' @name show_sig_bootstrap
 #' @return a `ggplot` object
@@ -55,25 +58,26 @@
 #'   ## bt_result = sig_fit_bootstrap_batch(mat, sig = mt_sig, n = 10, use_parallel = TRUE)
 #'
 #'   ## Show bootstrap exposure (optimal exposure is shown as triangle)
-#'   p1 <- show_sig_bootstrap_exposure(bt_result, methods = c("LS", "QP"))
+#'   p1 <- show_sig_bootstrap_exposure(bt_result, methods = c("QP"))
 #'   p1
-#'   p2 <- show_sig_bootstrap_exposure(bt_result, methods = c("LS", "QP"),
-#'                                                sample = "TCGA-AB-3012",
-#'                                                signatures = c("Sig1", "Sig2"))
+#'   p2 <- show_sig_bootstrap_exposure(bt_result,
+#'     methods = c("QP"),
+#'     sample = "TCGA-AB-3012",
+#'     signatures = c("Sig1", "Sig2")
+#'   )
 #'   p2
 #'
 #'   ## Show bootstrap error
-#'   p3 <- show_sig_bootstrap_error(bt_result, methods = c("LS", "QP"))
+#'   p3 <- show_sig_bootstrap_error(bt_result, methods = c("QP"))
 #'   p3
 #'
 #'   ## Show exposure (in)stability
-#'   p4 <- show_sig_bootstrap_stability(bt_result, methods = c("LS", "QP"))
+#'   p4 <- show_sig_bootstrap_stability(bt_result, methods = c("QP"))
 #'   p4
-#'   p5 <- show_sig_bootstrap_stability(bt_result, methods = c("LS", "QP"), measure = "MAE")
+#'   p5 <- show_sig_bootstrap_stability(bt_result, methods = c("QP"), measure = "MAE")
 #'   p5
-#'   p6 <- show_sig_bootstrap_stability(bt_result, methods = c("LS", "QP"), measure = "AbsDiff")
+#'   p6 <- show_sig_bootstrap_stability(bt_result, methods = c("QP"), measure = "AbsDiff")
 #'   p6
-#'
 #' } else {
 #'   message("Please install package 'BSgenome.Hsapiens.UCSC.hg19' firstly!")
 #' }
@@ -90,12 +94,20 @@ NULL
 #' @rdname show_sig_bootstrap
 #' @export
 show_sig_bootstrap_exposure <- function(bt_result, sample = NULL, signatures = NULL,
-                                        methods = "LS", palette = "aaas", title = NULL,
+                                        methods = "QP", plot_fun = c("boxplot", "violin"),
+                                        highlight = "auto",
+                                        palette = "aaas", title = NULL,
                                         xlab = FALSE, ylab = "Signature exposure", width = 0.3,
                                         dodge_width = 0.8, outlier.shape = NA,
                                         add = "jitter", add.params = list(alpha = 0.3),
                                         ...) {
   stopifnot(is.list(bt_result))
+  plot_fun <- match.arg(plot_fun)
+  plot_fun <- switch(
+    plot_fun,
+    boxplot = ggpubr::ggboxplot,
+    violin = ggpubr::ggviolin
+  )
 
   timer <- Sys.time()
   send_info("Started.")
@@ -126,29 +138,49 @@ show_sig_bootstrap_exposure <- function(bt_result, sample = NULL, signatures = N
 
   send_info("Plotting.")
   ## Plotting
-  ggpubr::ggboxplot(subset(dat, dat$type != "optimal"),
+  p <- plot_fun(subset(dat, dat$type != "optimal"),
     x = "sig", y = "exposure", color = "method", outlier.shape = outlier.shape,
     palette = palette, width = width, add = add, add.params = add.params,
     title = title, xlab = xlab, ylab = ylab, ...
-  ) +
-    ggplot2::geom_point(
+  )
+
+  if (highlight == "auto") {
+    p <- p + ggplot2::geom_point(
       data = subset(dat, dat$type == "optimal"),
-      mapping = aes_string(x = "sig", y = "exposure", color = "method"),
-      shape = 2, position = ggplot2::position_dodge2(width = dodge_width, preserve = "single")
+      mapping = ggplot2::aes_string(x = "sig", y = "exposure", color = "method"),
+      shape = 17, size = 4,
+      position = ggplot2::position_dodge2(width = dodge_width, preserve = "single")
     )
+  } else {
+    p <- p + ggplot2::geom_point(
+      data = subset(dat, dat$type == "optimal"),
+      mapping = ggplot2::aes_string(x = "sig", y = "exposure"),
+      shape = 17, size = 4, color = highlight,
+      position = ggplot2::position_dodge2(width = dodge_width, preserve = "single")
+    )
+  }
+  p
 }
 
 
 #' @rdname show_sig_bootstrap
 #' @export
 show_sig_bootstrap_error <- function(bt_result, sample = NULL,
-                                     methods = "LS", palette = "aaas", title = NULL,
-                                     xlab = FALSE, ylab = "Decomposition error", width = 0.3,
+                                     methods = "QP", plot_fun = c("boxplot", "violin"),
+                                     highlight = "auto",
+                                     palette = "aaas", title = NULL,
+                                     xlab = FALSE, ylab = "Reconstruction error (F2 norm)", width = 0.3,
                                      dodge_width = 0.8, outlier.shape = NA,
                                      add = "jitter", add.params = list(alpha = 0.3),
                                      legend = "none",
                                      ...) {
   stopifnot(is.list(bt_result))
+  plot_fun <- match.arg(plot_fun)
+  plot_fun <- switch(
+    plot_fun,
+    boxplot = ggpubr::ggboxplot,
+    violin = ggpubr::ggviolin
+  )
 
   timer <- Sys.time()
   send_info("Started.")
@@ -176,28 +208,47 @@ show_sig_bootstrap_error <- function(bt_result, sample = NULL,
 
   send_info("Plotting.")
   ## Plotting
-  ggpubr::ggboxplot(subset(dat, dat$type != "optimal"),
+  p <- plot_fun(subset(dat, dat$type != "optimal"),
     x = "method", y = "errors", color = "method", outlier.shape = outlier.shape,
     palette = palette, width = width, add = add, add.params = list(alpha = 0.3),
     title = title, xlab = xlab, ylab = ylab, legend = legend, ...
-  ) +
-    ggplot2::geom_point(
+  )
+
+  if (highlight == "auto") {
+    p <- p + ggplot2::geom_point(
       data = subset(dat, dat$type == "optimal"),
-      mapping = aes_string(x = "method", y = "errors", color = "method"),
-      shape = 2, position = ggplot2::position_dodge2(width = dodge_width, preserve = "single")
+      mapping = ggplot2::aes_string(x = "method", y = "errors", color = "method"),
+      shape = 17, size = 4,
+      position = ggplot2::position_dodge2(width = dodge_width, preserve = "single")
     )
+  } else {
+    p <- p + ggplot2::geom_point(
+      data = subset(dat, dat$type == "optimal"),
+      mapping = ggplot2::aes_string(x = "method", y = "errors"),
+      shape = 17, size = 4, color = highlight,
+      position = ggplot2::position_dodge2(width = dodge_width, preserve = "single")
+    )
+  }
+  p
 }
 
 
 #' @rdname show_sig_bootstrap
 #' @export
 show_sig_bootstrap_stability <- function(bt_result, signatures = NULL, measure = c("MRSE", "MAE", "AbsDiff"),
-                                         methods = "LS", palette = "aaas", title = NULL,
+                                         methods = "QP", plot_fun = c("boxplot", "violin"),
+                                         palette = "aaas", title = NULL,
                                          xlab = FALSE, ylab = "Signature instability",
                                          width = 0.3, outlier.shape = NA,
                                          add = "jitter", add.params = list(alpha = 0.3),
                                          ...) {
   stopifnot(is.list(bt_result))
+  plot_fun <- match.arg(plot_fun)
+  plot_fun <- switch(
+    plot_fun,
+    boxplot = ggpubr::ggboxplot,
+    violin = ggpubr::ggviolin
+  )
   measure <- match.arg(measure)
 
   timer <- Sys.time()
@@ -253,7 +304,7 @@ show_sig_bootstrap_stability <- function(bt_result, signatures = NULL, measure =
 
   send_info("Plotting.")
   ## Plotting
-  ggpubr::ggboxplot(dat,
+  plot_fun(dat,
     x = "sig", y = "measure", color = "method", outlier.shape = outlier.shape,
     palette = palette, width = width, add = add, add.params = add.params,
     title = title, xlab = xlab, ylab = ylab, ...